chore(decoder): move lvgl binary image decoder to standalone file
Binary image decoder is getting complex, it should be regarded as an implementation of image decoder.
